sqlmodel: move shared test helpers
